+++ client/src/components/ContentTitle.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
+import "../css/Contents.css";
 
 const ContentTitle = () => {
   const buttonStyle = {
