+++ client/src/components/Navigation.js
@@ -1,0 +1,11 @@
+import React from "react";
+import "../css/Navigation.css";
+
+const Navigation = () => {
+  return (
+    <div className="black-nav">
+      <h2> 개발 Blog </h2>
+    </div>
+  );
+};
+export default Navigation;
