+++ client/src/css/App.css
@@ -26,38 +26,9 @@
 body {
 	font-family : 'nanumsquare';
 }
-.black-nav {
-	background : black;
-	width : 100%;
-	display : flex;
-	color: white;
-	padding : 20px;
-	font-weight : 600;
-	font-size : 20px;
-}
 /*
 https://flexbox.help/
 */
-.contents_title {
-	display : flex;
-	padding-left : 60px;
-	text-align : left;
-	align-items : baseline;
-}
-.viewcontent {
-	display : flex;
-	justify-content : center;
-	align-items : space-around;
-	margin-top : 0px;
-}
-.contents {
-	flex-direction : column;
-	align-items : stretch; 
-	text-align : left;
-	line-height : 160%;
-	padding-left : 30px;
-	padding-top : 5px;
-}
 
 .App-link {
   color: #61dafb;
