+++ client/src/css/Contents.css
@@ -1,0 +1,21 @@
+
+.contents_title {
+	display : flex;
+	padding-left : 60px;
+	text-align : left;
+	align-items : baseline;
+}
+.viewcontent {
+	display : flex;
+	justify-content : center;
+	align-items : space-around;
+	margin-top : 0px;
+}
+.contents {
+	flex-direction : column;
+	align-items : stretch; 
+	text-align : left;
+	line-height : 160%;
+	padding-left : 30px;
+	padding-top : 5px;
+}
