+++ client/src/css/Navigation.css
@@ -1,0 +1,9 @@
+.black-nav {
+	background : black;
+	width : 100%;
+	display : flex;
+	color: white;
+	padding : 20px;
+	font-weight : 600;
+	font-size : 20px;
+}
